Use double precision for precision

+++ src/xmipp/libraries/reconstruction_cuda/cuda_forward_art_zernike3d.h
@@ -40,7 +40,7 @@ struct BlockSizes {
 	size_t x, y, z;
 };
 
-template<typename PrecisionType = float>
+template<typename PrecisionType = double>
 class Program {
 	static_assert(std::is_floating_point<PrecisionType>::value, "Floating point type is required.");
